requirements.txt
fastapi
uvicorn
python-multipart
deepface
supabase
pydantic
python-dotenv
pytest
httpx
opencv-python
numpy
